auth_service/requirements.txt
fastapi==0.101.1
uvicorn[standard]==0.23.1
pydantic[email]==2.0.3
sqlalchemy==2.0.20
asyncpg==0.27.0
pydantic-settings==2.0.3
alembic==1.12.0
pyjwt[crypto]==2.8.0
bcrypt==4.1.1
python-multipart==0.0.6
psycopg2-binary==2.9.6
aio-pika>=9.0.0